readme: copy walking-through example to tests
It can be useful to have a simple example with go-connector usage with
detailed comments. Walking-through example is a single example that left
in a README. The goal of example is to be an entrypoint for new users
and to be an initial snippet suitable for playing with Go connector.

Add a description with steps how-to run example manually.