readme: move section about testing to contributing guide
There is a guide made by Github with recommendations how to make project
healthy for contributors [1]. One of them is a contributing guide with
steps for new contributors. Knowledge about running tests in a project
is not needed to all people who will read our README, so I moved this
section to contributing guide.

1. https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions